apps/bot/requirements.txt
aiogram>=3.0.0
pydantic-settings
python-dotenv
httpx
sqlalchemy
psycopg2-binary
